backend/requirements.txt
fastapi
uvicorn
google-generativeai
firebase-admin
python-dotenv
pydantic
httpx
